+++ CHANGELOG.md
@@ -1,3 +1,53 @@
+## 3.40.0
+
+### 18 new icons:
+
+- `outline/api-book`
+- `outline/brand-infakt`
+- `outline/circle-open-arrow-down`
+- `outline/circle-open-arrow-left`
+- `outline/circle-open-arrow-right`
+- `outline/circle-open-arrow-up`
+- `outline/circle-plus-minus`
+- `outline/home-lock`
+- `outline/map-lock`
+- `outline/map-shield`
+- `outline/mesh`
+- `outline/queue-pop-in`
+- `outline/queue-pop-out`
+- `outline/stroke-curved`
+- `outline/stroke-dynamic`
+- `outline/stroke-straight`
+- `outline/trending-up-down`
+- `outline/vip-2`
+
+Fixed icons: `outline/api`, `outline/home-shield`
+
+## 3.39.0
+
+### 18 new icons:
+
+- `outline/ai-agent`
+- `outline/ai-agents`
+- `outline/ai-gateway`
+- `outline/bot-id`
+- `outline/canary`
+- `outline/codeblock`
+- `outline/connection`
+- `outline/credits`
+- `outline/firewall-check`
+- `outline/firewall-flame`
+- `outline/image-generation`
+- `outline/microfrontends`
+- `outline/middleware`
+- `outline/scan-traces`
+- `outline/serverless`
+- `outline/user-key`
+- `outline/world-map`
+- `outline/zero-config`
+
+Fixed icons: `outline/api`, `outline/sort-ascending-2`, `outline/sort-descending-2`
+
 ## 3.38.0
 
 ### 34 new icons: